Support running a different Kelvin benchmark binary
Allow passing in of 'kelvin_benchmark_info' which is a dictionary
that contains the needed info for the kelvin_binary (e.g srcs).
This allows the benchmark to run a different Kelvin binary while
still keeping the SEC and SMC ones the same.

Tested:
best_of_times, silence with ISS+FPGA

Change-Id: Iabba3fd2e19524a66c3f530f095828e5201d0e9c