Rewrite ConvPerChannelD1OD24_5x5
This new approach utilizes the aconv instruction to perform full matmul.
Creative data shuffling methods allow the implementation to achieve
significantly better useful MAC throughput.

Activation min/max is assumed to be int8 limits to eliminate the clamping
instructions in the postprocessing

ambient_face_2.8.0_504x378_grayscale_per_axis_simulator_iss is used in
testing.
The following items are measured on the simulator:
- Mismatch_count is 0.
- #0 CONV_2D takes 1.52M ticks (down from 4.02M, 2.64x speedup).
- Total cycles is now 52.9M (down from 57.9M, -8.6%).

The implementation currently contains lots of repeated and near-repeated
sections. They can be folded together in the future to improve
readability.

The following improvement items may be doable:
- Further tweak data shuffling sections to take less time
- Rearrange instructions on bottlenecks

Change-Id: Ibfcea94988b9ece4ea98a61c139ce44e967b4fd8